Introduce custom Chat component
This component should allow clients to set their own custom Emoji
components and dataset. With tree-shaking, this helps to have more
control over what they want to use from the library or not.

The component could be later extended to provide other types of
customizations that rely on external packages that other wise could be
dropped with tree-shaking.